requirements.txt
google-generativeai==0.7.0
python-dotenv==1.0.0

